created useEnterKeyPress as a helper hook
